to int

+++ src/drawing_utils.py
@@ -4,7 +4,7 @@ import cv2
 
 
 def draw_label(I,l,color=(255,0,0),thickness=1):
-	wh = np.array(I.shape[1::-1]).astype(float)
+	wh = np.array(I.shape[1::-1]).astype(int)
 	tl = tuple((l.tl()*wh).astype(int).tolist())
 	br = tuple((l.br()*wh).astype(int).tolist())
 	cv2.rectangle(I,tl,br,color,thickness=thickness)
